BCP-39 #done #time 5m BCP-40 #done BCP-41 #done #time 4h
Cytoscape has a task called CreateNetworkViewTask which does this without bugs. This task can be created with the CreateNetworkViewTaskFactory service, and executed by the main TaskManager. This creates a new CyNetworkView using the default style and layout.

Network refreshing is now archieved by a workaround, by destroying all current views for the network, and recreating them in the manner above. This might not be a long-term solution.