BGR-39 #done #time 3h
Added support for searching for PPIs with a minimum number of common connections.

Also added support for extra columns in the BGRelationSearchResultView. BGRelation has a new extraTableData ArrayList where any data can be added.

BGRelationSearchResultView now gets a sampleRow at init, and defines a DefaultTableModel which uses the sampleRow to define the correct data type for sorting.